refactor: PostManagerPage 컴포넌트 분리
