+++ src/pages/types/postsManagerTypes.ts
