+++ src/pages/ui/comment/CommentAddButton.tsx
@@ -1,0 +1,16 @@
+import { Button } from "../../../shared/ui"
+
+export const CommentAddButton = () => {
+  return (
+    <Button
+      size="sm"
+      onClick={() => {
+        setNewComment((prev) => ({ ...prev, postId }))
+        setShowAddCommentDialog(true)
+      }}
+    >
+      <Plus className="w-3 h-3 mr-1" />
+      댓글 추가
+    </Button>
+  )
+}
