+++ src/pages/ui/comment/CommentAddDialog.tsx
@@ -1,0 +1,20 @@
+export const CommentAddDialog = () => {
+  return (
+    // {/* 댓글 추가 대화상자  entity/comments/ui/comment-add-dialog.tsx*/}
+    <Dialog open={showAddCommentDialog} onOpenChange={setShowAddCommentDialog}>
+      <DialogContent>
+        <DialogHeader>
+          <DialogTitle>새 댓글 추가</DialogTitle>
+        </DialogHeader>
+        <div className="space-y-4">
+          <Textarea
+            placeholder="댓글 내용"
+            value={newComment.body}
+            onChange={(e) => setNewComment({ ...newComment, body: e.target.value })}
+          />
+          <Button onClick={addComment}>댓글 추가</Button>
+        </div>
+      </DialogContent>
+    </Dialog>
+  )
+}
