+++ src/pages/ui/comment/CommentEditButton.tsx
@@ -1,0 +1,16 @@
+import { Button } from "../../../shared/ui"
+
+export const CommentEditButton = () => {
+  return (
+    <Button
+      variant="ghost"
+      size="sm"
+      onClick={() => {
+        setSelectedComment(comment)
+        setShowEditCommentDialog(true)
+      }}
+    >
+      <Edit2 className="w-3 h-3" />
+    </Button>
+  )
+}
