+++ src/pages/ui/comment/CommentEditDialog.tsx
@@ -1,0 +1,20 @@
+export const CommentEditDialog = () => {
+  return (
+    // {/* 댓글 수정 대화상자  entity/comments/ui/comment-edit-dialog.tsx*/}
+    <Dialog open={showEditCommentDialog} onOpenChange={setShowEditCommentDialog}>
+      <DialogContent>
+        <DialogHeader>
+          <DialogTitle>댓글 수정</DialogTitle>
+        </DialogHeader>
+        <div className="space-y-4">
+          <Textarea
+            placeholder="댓글 내용"
+            value={selectedComment?.body || ""}
+            onChange={(e) => setSelectedComment({ ...selectedComment, body: e.target.value })}
+          />
+          <Button onClick={updateComment}>댓글 업데이트</Button>
+        </div>
+      </DialogContent>
+    </Dialog>
+  )
+}
