+++ src/pages/ui/comment/CommentLikeButton.tsx
@@ -1,0 +1,8 @@
+export const CommentLikeButton = () => {
+  return (
+    <Button variant="ghost" size="sm" onClick={() => likeComment(comment.id, postId)}>
+      <ThumbsUp className="w-3 h-3" />
+      <span className="ml-1 text-xs">{comment.likes}</span>
+    </Button>
+  )
+}
