+++ src/pages/ui/comment/CommentList.tsx
@@ -1,0 +1,41 @@
+import { Loading } from "../common/Loading"
+import { CommentAddButton } from "./CommentAddButton"
+import { CommentDeleteButton } from "./CommentDeleteButton"
+import { CommentEditButton } from "./CommentEditButton"
+import { CommentLikeButton } from "./CommentLikeButton"
+
+export const Comments = () => {
+  const { isLoading, comments, error } = useComments()
+
+  if (isLoading) {
+    return <Loading />
+  }
+
+  if (error) {
+    return <div className="mt-2 text-sm text-red-500">댓글을 불러오는 중 오류가 발생했습니다.</div>
+  }
+
+  return (
+    <div className="mt-2">
+      <div className="flex items-center justify-between mb-2">
+        <h3 className="text-sm font-semibold">댓글</h3>
+        <CommentAddButton />
+      </div>
+      <div className="space-y-1">
+        {comments[postId]?.map((comment) => (
+          <div key={comment.id} className="flex items-center justify-between text-sm border-b pb-1">
+            <div className="flex items-center space-x-2 overflow-hidden">
+              <span className="font-medium truncate">{comment.user.username}:</span>
+              <span className="truncate">{highlightText(comment.body, searchQuery)}</span>
+            </div>
+            <div className="flex items-center space-x-1">
+              <CommentLikeButton />
+              <CommentEditButton />
+              <CommentDeleteButton />
+            </div>
+          </div>
+        ))}
+      </div>
+    </div>
+  )
+}
