+++ src/pages/ui/common/Loading.tsx
@@ -1,0 +1,3 @@
+export const Loading = () => {
+  return <div className="flex justify-center p-4">로딩 중...</div>
+}
