+++ src/pages/ui/post/PostAddDialog.tsx
@@ -1,0 +1,40 @@
+import { DialogHeader, DialogTitle, Input } from "../../../shared/ui"
+
+import { Button } from "../../../shared/ui"
+
+import { DialogContent } from "../../../shared/ui"
+
+import { Dialog } from "../../../shared/ui"
+
+export const PostAddDialog = () => {
+  return (
+    // {/* 게시물 추가 대화상자  entity/posts/ui/post-add-dialog.tsx*/}
+    <Dialog open={showAddDialog} onOpenChange={setShowAddDialog}>
+      <DialogContent>
+        <DialogHeader>
+          <DialogTitle>새 게시물 추가</DialogTitle>
+        </DialogHeader>
+        <div className="space-y-4">
+          <Input
+            placeholder="제목"
+            value={newPost.title}
+            onChange={(e) => setNewPost({ ...newPost, title: e.target.value })}
+          />
+          <Textarea
+            rows={30}
+            placeholder="내용"
+            value={newPost.body}
+            onChange={(e) => setNewPost({ ...newPost, body: e.target.value })}
+          />
+          <Input
+            type="number"
+            placeholder="사용자 ID"
+            value={newPost.userId}
+            onChange={(e) => setNewPost({ ...newPost, userId: Number(e.target.value) })}
+          />
+          <Button onClick={addPost}>게시물 추가</Button>
+        </div>
+      </DialogContent>
+    </Dialog>
+  )
+}
