+++ src/pages/ui/post/PostDetail.tsx
@@ -1,0 +1,20 @@
+import { DialogHeader, DialogTitle } from "../../../shared/ui"
+
+import { Dialog } from "../../../shared/ui"
+import { highlightText } from "../../utils/highlightText"
+
+export const PostDetail = () => {
+  return (
+    <Dialog open={showPostDetailDialog} onOpenChange={setShowPostDetailDialog}>
+      <DialogContent className="max-w-3xl">
+        <DialogHeader>
+          <DialogTitle>{highlightText(selectedPost?.title, searchQuery)}</DialogTitle>
+        </DialogHeader>
+        <div className="space-y-4">
+          <p>{highlightText(selectedPost?.body, searchQuery)}</p>
+          {renderComments(selectedPost?.id)}
+        </div>
+      </DialogContent>
+    </Dialog>
+  )
+}
