+++ src/pages/ui/post/PostEditDialog.tsx
@@ -1,0 +1,36 @@
+import { Button } from "../../../shared/ui"
+
+import { Input, Textarea } from "../../../shared/ui"
+
+import { DialogHeader, DialogTitle } from "../../../shared/ui"
+
+import { DialogContent } from "../../../shared/ui"
+
+import { Dialog } from "../../../shared/ui"
+
+export const PostEditDialog = () => {
+  return (
+    // {/* 게시물 수정 대화상자  entity/posts/ui/post-edit-dialog.tsx*/}
+    <Dialog open={showEditDialog} onOpenChange={setShowEditDialog}>
+      <DialogContent>
+        <DialogHeader>
+          <DialogTitle>게시물 수정</DialogTitle>
+        </DialogHeader>
+        <div className="space-y-4">
+          <Input
+            placeholder="제목"
+            value={selectedPost?.title || ""}
+            onChange={(e) => setSelectedPost({ ...selectedPost, title: e.target.value })}
+          />
+          <Textarea
+            rows={15}
+            placeholder="내용"
+            value={selectedPost?.body || ""}
+            onChange={(e) => setSelectedPost({ ...selectedPost, body: e.target.value })}
+          />
+          <Button onClick={updatePost}>게시물 업데이트</Button>
+        </div>
+      </DialogContent>
+    </Dialog>
+  )
+}
