+++ src/pages/ui/post/PostList.tsx
@@ -1,0 +1,26 @@
+import { useState } from "react"
+import { Pagination } from "../common/Pagination"
+import { PostSearch } from "./PostSearch"
+import { PostTable } from "./PostTable"
+import { Loading } from "../common/Loading"
+
+export const PostList = () => {
+  // 얘는 추후 탄스택 쿼리로 변경
+  const [loading, setLoading] = useState(false)
+
+  // 이 컴포넌트에서는 query를 공유해야한다.
+  // const { data, isLoading } = useGetPostsQuery(query)
+
+  return (
+    <div className="flex flex-col gap-4">
+      {/* 검색 및 필터 컨트롤 */}
+      <PostSearch />
+
+      {/* 게시물 테이블 */}
+      {loading ? <Loading /> : <PostTable />}
+
+      {/* 페이지네이션 */}
+      <Pagination limit={limit} setLimit={setLimit} skip={skip} setSkip={setSkip} total={total} />
+    </div>
+  )
+}
