+++ src/pages/ui/post/index.ts
@@ -1,0 +1,8 @@
+export * from "./PostAddDialog"
+export * from "./PostEditDialog"
+export * from "./PostList"
+export * from "./PostSearch"
+export * from "./PostTable"
+export * from "./PostManagerLayout"
+export * from "./PostHeader"
+export * from "./PostDetail"
