+++ src/shared/ui/Card.tsx
@@ -1,22 +1,28 @@
-import { forwardRef } from "react"
+import { forwardRef, PropsWithChildren } from "react"
 
 // 카드 컴포넌트
-export const Card = forwardRef<HTMLDivElement, { className?: string }>(({ className, ...props }, ref) => (
-  <div ref={ref} className={`rounded-lg border bg-card text-card-foreground shadow-sm ${className}`} {...props} />
-))
+export const Card = forwardRef<HTMLDivElement, PropsWithChildren<{ className?: string }>>(
+  ({ className, ...props }, ref) => (
+    <div ref={ref} className={`rounded-lg border bg-card text-card-foreground shadow-sm ${className}`} {...props} />
+  ),
+)
 Card.displayName = "Card"
 
-export const CardHeader = forwardRef<HTMLDivElement, { className?: string }>(({ className, ...props }, ref) => (
-  <div ref={ref} className={`flex flex-col space-y-1.5 p-6 ${className}`} {...props} />
-))
+export const CardHeader = forwardRef<HTMLDivElement, PropsWithChildren<{ className?: string }>>(
+  ({ className, ...props }, ref) => (
+    <div ref={ref} className={`flex flex-col space-y-1.5 p-6 ${className}`} {...props} />
+  ),
+)
 CardHeader.displayName = "CardHeader"
 
-export const CardTitle = forwardRef<HTMLHeadingElement, { className?: string }>(({ className, ...props }, ref) => (
-  <h3 ref={ref} className={`text-2xl font-semibold leading-none tracking-tight ${className}`} {...props} />
-))
+export const CardTitle = forwardRef<HTMLHeadingElement, PropsWithChildren<{ className?: string }>>(
+  ({ className, ...props }, ref) => (
+    <h3 ref={ref} className={`text-2xl font-semibold leading-none tracking-tight ${className}`} {...props} />
+  ),
+)
 CardTitle.displayName = "CardTitle"
 
-export const CardContent = forwardRef<HTMLDivElement, { className?: string }>(({ className, ...props }, ref) => (
-  <div ref={ref} className={`p-6 pt-0 ${className}`} {...props} />
-))
+export const CardContent = forwardRef<HTMLDivElement, PropsWithChildren<{ className?: string }>>(
+  ({ className, ...props }, ref) => <div ref={ref} className={`p-6 pt-0 ${className}`} {...props} />,
+)
 CardContent.displayName = "CardContent"
